Move redux aware stuff into own folder
