+++ src/index.js
@@ -5,7 +5,7 @@ import {render} from 'react-dom';
 import { Provider } from 'react-redux';
 import { Router, browserHistory } from 'react-router';
 import routes from './routes';
-import configureStore from './store/configureStore';
+import configureStore from './redux/store/configureStore';
 require('./favicon.ico'); // Tell webpack to load favicon.ico
 
 import { syncHistoryWithStore } from 'react-router-redux';
