+++ src/redux/reducers/appReducer.spec.js
