+++ src/routes.js
@@ -2,8 +2,8 @@ import React from 'react';
 import {Route, IndexRoute} from 'react-router';
 
 import App from './components/App';
-import SearchPage from './components/SearchPage';
-import ArtistDetails from './components/ArtistDetails';
+import SearchPage from './redux/root-containers/SearchPage';
+import ArtistDetails from './redux/root-containers/ArtistDetails';
 import NotFoundPage from './components/NotFoundPage.js';
